+++ bootstrap/app_conversion.go
@@ -20,28 +20,6 @@ func appHardwareWrapper[THardware europi.Hardware](app any) any {
 	}
 }
 
-func getAppFuncs(e europi.Hardware, app any) (start AppStartFunc, mainLoop AppMainLoopFunc, end AppEndFunc) {
-	if appStart, _ := app.(ApplicationStart[europi.Hardware]); appStart != nil {
-		start = appStart.Start
-	}
-	if appMainLoop, _ := app.(ApplicationMainLoop[europi.Hardware]); appMainLoop != nil {
-		mainLoop = appMainLoop.MainLoop
-	}
-	if appEnd, _ := app.(ApplicationEnd[europi.Hardware]); appEnd != nil {
-		end = appEnd.End
-	}
-
-	switch e.(type) {
-	case *europi.EuroPiPrototype:
-		start, mainLoop, end = getWrappedAppFuncs[*europi.EuroPiPrototype](app)
-	case *europi.EuroPi:
-		start, mainLoop, end = getWrappedAppFuncs[*europi.EuroPi](app)
-		// TODO: add rev2
-	}
-
-	return
-}
-
 func getWrappedAppFuncs[THardware europi.Hardware](app any) (start AppStartFunc, mainLoop AppMainLoopFunc, end AppEndFunc) {
 	appWrapper := appHardwareWrapper[THardware](app)
 	if getStart, _ := appWrapper.(applicationStartProvider); getStart != nil {
