+++ internal/nonpico/nonpico.go
@@ -1,44 +1,7 @@
-//go:build !pico
-// +build !pico
-
 package nonpico
 
-import (
-	"github.com/awonak/EuroPiGo/hardware"
-	"github.com/awonak/EuroPiGo/hardware/hal"
-	"github.com/awonak/EuroPiGo/internal/nonpico/rev0"
-	"github.com/awonak/EuroPiGo/internal/nonpico/rev1"
-)
-
-func initRevision0() {
-	rev0.DoInit()
-}
-
-func initRevision1() {
-	rev1.DoInit()
-}
-
-func initRevision2() {
-	//TODO: rev2.DoInit()
-}
-
-var detectedRevision hal.Revision
-
-func DetectRevision() hal.Revision {
-	return detectedRevision
-}
+// This file is required for compilation to occur. Do not remove it
+// or the empty init function at the bottom of the file.
 
 func init() {
-	hardware.OnRevisionDetected(func(revision hal.Revision) {
-		switch revision {
-		case hal.Revision0:
-			initRevision0()
-		case hal.Revision1:
-			initRevision1()
-		case hal.Revision2:
-			initRevision2()
-		default:
-		}
-		hardware.SetReady()
-	})
 }
