--- revisiondetection.go
@@ -1,7 +1,0 @@
-package europi
-
-var ensureHardware func()
-
-func EnsureHardware() {
-	ensureHardware()
-}
